fix updating utxos multiple times
